need to fix data types for loop indices

+++ cuda3.cu
@@ -30,7 +30,8 @@ double r8_uniform_01 ( int *seed ){
   return r;
 }
 
-//We will use this define statement to specify what the size of the thread blocks should be 
+//We will use this define statement to specify what the size of the thread blocks should be
+//optimum has been determined to be 16
 #define BLOCK_SIZE 16
 
 __global__ void unoptimized_mult_kernel(float* a, float* b, float* c, unsigned long long dimension) {
@@ -50,8 +51,6 @@ accumulator += (a[row * dimension + e]) * (b[e * dimension + col]);
 c[row * dimension + col] = accumulator;
 }
 
-// Matrix multiplication - Host code
-// Matrix dimensions are assumed to be multiples of BLOCK_SIZE
 
 double* unoptimized_mult( float* a, float* b, float* c, int run_number) {
 
@@ -65,7 +64,9 @@ unsigned long long n = dimension;
 
 // Load A and B to device memory
 float* cuda_A;
-size_t size = l * l * sizeof(float);
+//size_t size = l * l * sizeof(float);
+unsigned long long size = l*l*sizeof(float);
+
 cudaMalloc(&cuda_A, size);
 cudaMemcpy(cuda_A, a, size, cudaMemcpyHostToDevice);
 
@@ -118,7 +119,7 @@ cudaThreadSynchronize();
 //CUDA timekeeping stop events
 cudaEventRecord(stop,0);
 cudaEventSynchronize(stop);
-//store the elapse dtime
+//store the elapsed time
 cudaEventElapsedTime(&time_elapsed,start,stop);
 cudaEventDestroy(start);
 cudaEventDestroy(stop);
@@ -127,7 +128,7 @@ cudaEventDestroy(stop);
 unsigned long long ops = l * l * ( 2 * l );
 
 time_elapsed = time_elapsed/1000; //change into seconds
-//figure out the rate  
+//figure out the rate
 double rate = ( double ) ( ops ) / time_elapsed / 1000000.0;
 
   printf ( "\n" );
@@ -142,11 +143,11 @@ cudaMemcpy(c, cuda_C, size, cudaMemcpyDeviceToHost);
 
 
 
-
+//return values for main
   dt_and_rate[0]=time_elapsed;
   dt_and_rate[1]=rate;
 
-  
+
 
 
 
@@ -160,26 +161,22 @@ return dt_and_rate;
 
 
 
-// Get a matrix element
+//retrieve a matrix element
+//device function because this needs to execute on the graphics card, not on host machines
 __device__ float retrieve_item(float* a, int row, int col, unsigned long long dim) {
 return a[row * dim + col];
 }
 
 
 // Set a matrix element
+//device function because this needs to execute on the graphics card, not on host machines
 __device__ void set_entry(float* a, int row, int col, float value, unsigned long long dim) {
 a[row * dim + col] = value;
 }
 
-// Get the BLOCK_SIZExBLOCK_SIZE sub-matrix Asub of A that is
-
-// located col sub-matrices to the right and row sub-matrices down
-// from the upper-left corner of A
+//get the next sub-block of a matrix when doing blocking multiplication
 __device__ float* get_next_block(float* a, int row, int col, unsigned long long dimension) {
 float* block_a;
-//Asub.width = BLOCK_SIZE;
-//Asub.height = BLOCK_SIZE;
-//Asub.stride = A.stride;
 block_a = &a[dimension * BLOCK_SIZE * row + BLOCK_SIZE * col];
 return block_a;
 }
@@ -190,27 +187,26 @@ __global__ void blocking_mult_kernel(float* a, float* b, float* c, unsigned long
 // Block row and column
 int blockRow = blockIdx.y;
 int blockCol = blockIdx.x;
-// Each thread block computes one sub-matrix Csub of C
+// Each thread block computes one sub-matrix of C
 float* block_c = get_next_block(c, blockRow, blockCol, dimension);
-// Each thread computes one element of Csub
+// Each thread computes one element of the tile of C
 // by accumulating results into accumulator
 float accumulator = 0.0;
-// Thread row and column within Csub
+// Thread row and column within the tile of c
 int row = threadIdx.y;
 int col = threadIdx.x;
-// Loop over all the sub-matrices of A and B that are
-// required to compute Csub
+// Loop over all A and B's tiles
 // Multiply each pair of sub-matrices together
 // and accumulate the results
 for (int m = 0; m < (dimension / BLOCK_SIZE); ++m) {
-// Get sub-matrix Asub of A
+// Get sub-matrix of A
 float* block_a = get_next_block(a, blockRow, m, dimension);
-// Get sub-matrix Bsub of B
+// Get sub-matrix of B
 float* block_b = get_next_block(b, m, blockCol, dimension);
-// Shared memory used to store Asub and Bsub respectively
+// Shared memory used to store the tiles of A and B
 __shared__ float a_shared[BLOCK_SIZE][BLOCK_SIZE];
 __shared__ float b_shared[BLOCK_SIZE][BLOCK_SIZE];
-// Load Asub and Bsub from device memory to shared memory
+// Load tiles from device memory to shared memory
 // Each thread loads one element of each sub-matrix
 a_shared[row][col] = retrieve_item(block_a, row, col, dimension);
 b_shared[row][col] = retrieve_item(block_b, row, col, dimension);
@@ -225,7 +221,7 @@ accumulator += a_shared[row][e] * b_shared[e][col];
 // sub-matrices of A and B in the next iteration
 __syncthreads();
 }
-// Write Csub to device memory
+// Write the tile of to device memory
 // Each thread writes one element
 set_entry(block_c, row, col, accumulator, dimension);
 }
@@ -244,45 +240,38 @@ unsigned long long n = dimension;
 
 
 // Load A and B to device memory
-
 float* cuda_A;
 
 size_t size = dimension * dimension * sizeof(float);
+//allocate device memory
 cudaMalloc(&cuda_A, size);
-
+//copy to device
 cudaMemcpy(cuda_A, a, size, cudaMemcpyHostToDevice);
-
+//do the same for b
 float* cuda_B;
-
 size = dimension * dimension * sizeof(float);
 cudaMalloc(&cuda_B, size);
-
-
 cudaMemcpy(cuda_B, b, size, cudaMemcpyHostToDevice);
 // Allocate C in device memory
-//Matrix d_C;
 float* cuda_C;
-
 size = dimension * dimension * sizeof(float);
 cudaMalloc(&cuda_C, size);
-
-
+//set up float for timer
 float time_elapsed;
 
-//struct timeval t1, t2;
 
-//gettimeofday(&t1, 0);
-// Invoke kernel
+//st up the block and grid for device
 dim3 dimBlock(BLOCK_SIZE, BLOCK_SIZE);
 dim3 dimGrid(dimension / dimBlock.x, dimension / dimBlock.y);
-
+//start timers
 cudaEvent_t start, stop;
 cudaEventCreate(&start);
 cudaEventCreate(&stop);
 cudaEventRecord(start,0);
 
-
+//call the CUDA function
 blocking_mult_kernel<<<dimGrid, dimBlock>>>(cuda_A, cuda_B, cuda_C, dimension);
+//synchronize threads
 cudaThreadSynchronize();
 
 cudaEventRecord(stop,0);
@@ -290,11 +279,11 @@ cudaEventSynchronize(stop);
 cudaEventElapsedTime(&time_elapsed,start,stop);
 cudaEventDestroy(start);
 cudaEventDestroy(stop);
-
+//figure out how many operations we have done
 unsigned long long ops = l * l * ( 2 * l );
 
 time_elapsed = time_elapsed/1000; //change into seconds
-  
+  //get the rate
   double rate = ( double ) ( ops ) / time_elapsed / 1000000.0;
 
   printf ( "\n" );
@@ -305,7 +294,7 @@ time_elapsed = time_elapsed/1000; //change into seconds
 
 
 
-// Read C from device memory
+// Read c from device memory
 cudaMemcpy(c, cuda_C, size, cudaMemcpyDeviceToHost);
 
 // Free device memory
@@ -338,14 +327,14 @@ void usage (char* argv[])
     exit(0);
 }
 
-
+//main function
 int main(int argc, char* argv[]){
 
-
+//used to determine average rates and times
 double* temp;
 double average_dt=0.0;
 double average_rate=0.0;
-
+//get command line options
 char options;
     while( (options=getopt(argc,argv,"n:h")) != -1){
         switch(options){
@@ -360,7 +349,7 @@ char options;
             exit(1);
         }
     }
-
+//allocate space
 float* a = (float*)malloc(dimension * dimension * sizeof(float));
 
 float* b = (float*)malloc(dimension * dimension * sizeof(float));
@@ -369,17 +358,17 @@ float* c = (float*)malloc(dimension * dimension * sizeof(float));
 
 int seed=123456789;
 
-
+//seed the matrices with random values
 for(int i = 0; i < dimension; i++){
   for(int j = 0; j < dimension; j++){
-    
+
     a[i*dimension + j] = (float) (r8_uniform_01 ( &seed ));
   }
 }
-  
+
 for(int i = 0; i < dimension; i++){
   for(int j = 0; j < dimension; j++){
-  
+
     b[i*dimension + j] = (float) (r8_uniform_01 ( &seed ));
   }
 }
@@ -421,7 +410,7 @@ average_rate=average_dt=0.0;
 
 
 
-
+//free host memory
 free(a);
 free(b);
 free(c);
